Time: 2 ms (99.83%) | Memory: 69.7 MB (44.19%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.java
@@ -1,11 +1,11 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
     
-        double sum=0;
+        int sum=0;
         for(int i=0;i<k;i++){
             sum+=nums[i];
         }
-        double max=sum;
+        int max=sum;
         for(int i=k;i<nums.length;i++){
              sum = sum + nums[i] - nums[i - k];
